:recycle: refactor: add unexpected fields and fix types

+++ apps/checklistsD/models.py
@@ -59,7 +59,7 @@ class ChecklistD(models.Model):
 
 class Transformer(models.Model):
     checklist = models.ForeignKey(ChecklistD, on_delete=models.CASCADE, related_name="transformers")
-    power = models.PositiveIntegerField(_("Transformer Power"))
+    power = models.FloatField(_("Transformer Power"))
     impedance = models.FloatField(_("Impedance"), blank=True, null=True)
     demand = models.FloatField(_("Demand"), blank=True, null=True)
     type = models.CharField(_("Type"), choices=TYPE_CHOICES, max_length=3, blank=True, null=True)
